+++ R/Object_Utilities.R
@@ -183,7 +183,7 @@ Add_Mito_Ribo_Seurat <- function(
       )
     }
     cli_inform(message = c("Columns with {.val {mito_name}} and/or {.val {ribo_name}} already present in meta.data slot.",
-                           "i" = "Overwriting those columns as .code {overwrite = TRUE.}")
+                           "i" = "Overwriting those columns as {.code overwrite = TRUE.}")
     )
   }
 
@@ -391,9 +391,10 @@ Add_Cell_Complexity_Seurat <- function(
 #' storing corrected and uncorrected assays in same object (e.g. outputs of both Cell Ranger and Cell Bender).
 #' @param overwrite Logical.  Whether to overwrite existing an meta.data column.  Default is FALSE meaning that
 #' function will abort if column with name provided to `meta_col_name` is present in meta.data slot.
+#' @param verbose logical, whether to print messages with status updates, default is TRUE.
 #'
 #' @import cli
-#' @importFrom dplyr select all_of
+#' @importFrom dplyr select all_of bind_rows
 #' @importFrom magrittr "%>%"
 #' @importFrom rlang is_installed
 #' @importFrom SeuratObject LayerData
@@ -425,7 +426,8 @@ Add_Top_Gene_Pct_Seurat <- function(
     num_top_genes = 50,
     meta_col_name = NULL,
     assay = "RNA",
-    overwrite = FALSE
+    overwrite = FALSE,
+    verbose = TRUE
 ){
   # Check for scuttle first
   scuttle_check <- is_installed(pkg = "scuttle")
@@ -468,15 +470,47 @@ Add_Top_Gene_Pct_Seurat <- function(
     )
   }
 
+  count_layers_present <- Layers(object = seurat_object, search = "counts")
+
   # Extract matrix
-  count_mat <- LayerData(object = seurat_object, assay = assay)
+  if (length(x = count_layers_present) == 1) {
+    if (isTRUE(x = verbose)) {
+      cli_inform(message = "Calculating percent expressing top {num_top_genes} for layer: {.field {count_layers_present}}")
+    }
 
-  # calculate
-  res <- as.data.frame(scuttle::perCellQCMetrics(x = count_mat, percent.top = num_top_genes))
+    count_mat <- LayerData(object = seurat_object, assay = assay, layer = "counts")
 
-  # select percent column
-  res <- res %>%
-    select(all_of(scuttle_colname))
+    # calculate
+    res <- as.data.frame(scuttle::perCellQCMetrics(x = count_mat, percent.top = num_top_genes))
+
+    # select percent column
+    res <- res %>%
+      select(all_of(scuttle_colname))
+  }
+
+
+  if (length(x = count_layers_present) > 1) {
+    res_list <- lapply(1:length(x = count_layers_present), function(x) {
+      if (isTRUE(x = verbose)) {
+        cli_inform(message = "Calculating percent expressing top {num_top_genes} for layer: {.field {count_layers_present[x]}}")
+      }
+
+      # Get layer data
+      layer_count <- LayerData(object = seurat_object, assay = assay, layer = count_layers_present[x])
+
+      # run scuttle
+      layer_res <- as.data.frame(scuttle::perCellQCMetrics(x = layer_count, percent.top = num_top_genes))
+      # select results column
+      layer_res <- layer_res %>%
+        select(all_of(scuttle_colname))
+    })
+
+    # combine results
+    if (isTRUE(x = verbose)) {
+      cli_inform(message = "Combining data from: {.field {count_layers_present}}")
+    }
+    res <- bind_rows(res_list)
+  }
 
   # Add to object and return
   seurat_object <- AddMetaData(object = seurat_object, metadata = res, col.name = meta_col_name)